feat: extract TasksFilters client component
